feat: add dragonfly into the GO archs

+++ agent/fileutil/fileutil_unix.go
@@ -11,8 +11,8 @@
 // either express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-//go:build darwin || freebsd || linux || netbsd || openbsd
-// +build darwin freebsd linux netbsd openbsd
+//go:build darwin || freebsd || linux || netbsd || openbsd || dragonfly
+// +build darwin freebsd linux netbsd openbsd dragonfly
 
 // Package fileutil contains utilities for working with the file system.
 package fileutil
@@ -106,8 +106,8 @@ func GetDiskSpaceInfo() (diskSpaceInfo DiskSpaceInfo, err error) {
 	return DiskSpaceInfo{
 		// On Linux the struct statfs.f_bavail field is unsigned, but on FreeBSD the field is an int64
 		AvailBytes: (int64)((uint64)(stat.Bavail) * bSize), // available space = # of available blocks * block size
-		FreeBytes:  (int64)(stat.Bfree * bSize),            // free space = # of free blocks * block size
-		TotalBytes: (int64)(stat.Blocks * bSize),           // total space = # of total blocks * block size
+		FreeBytes:  (int64)((uint64)(stat.Bfree) * bSize),  // free space = # of free blocks * block size
+		TotalBytes: (int64)((uint64)(stat.Blocks) * bSize), // total space = # of total blocks * block size
 	}, nil
 }
